added the download file for crypto functions
I added it so we can write files for the eventual functions we are going to implement. what would be better is that if we could specify a path for the user so that they can obtain the file more obvious to see and add a noti that the file has been downloaded.